Remove monkeypatching for ES7.
