+++ docs/architecture/decisions/0003-es-aaq-documents.md
@@ -1,4 +1,4 @@
-# 3. AAQ structure in Search v2
+# 3. AAQ structure in Search
 
 Date: 2020-10-27
 
@@ -12,16 +12,17 @@ As we are re-implementing our search in ElasticSearch v7,
 we must re-implement Ask a Question (AAQ) search.
 
 There is one primary use-case for storing AAQ documents in ES which Search v1 supports,
-which we must continue to be able to do in Search v2:
+which we must continue to be able to do in the redesigned Search:
 searching for an AAQ thread as a unit.
 
 There are other secondary use-cases which we may want to support when storing AAQ documents in ES.
 A non-exhaustive list of these are:
-* Searching within AAQ threads
-* Searching within questions and their solutions
-* Aggregating answers to create contribution data
 
-We also want search to be *fast*,
+-   Searching within AAQ threads
+-   Searching within questions and their solutions
+-   Aggregating answers to create contribution data
+
+We also want search to be _fast_,
 so should model our data to avoid nested fields and parent-child relationships,
 and use de-normalization wherever possible:
 https://www.elastic.co/guide/en/elasticsearch/reference/7.9/tune-for-search-speed.html#_document_modeling
@@ -64,18 +65,18 @@ We will set this to null in the `AnswerDocument`.
 
 This structure gives us great flexibility over the type of searches and aggregations we can do:
 
-* AAQ threads can be searched over as a unit by searching over the `question_content` and `answer_content` fields on a `QuestionDocument`.
-Collapsing or aggregations aren't required to ensure that threads only appear once,
-which wouldn't be the case if we didn't have an `answer_content` field.
+-   AAQ threads can be searched over as a unit by searching over the `question_content` and `answer_content` fields on a `QuestionDocument`.
+    Collapsing or aggregations aren't required to ensure that threads only appear once,
+    which wouldn't be the case if we didn't have an `answer_content` field.
 
-* Searches within AAQ threads are possible by searching over the `question_content` fields on a `QuestionDocument`,
-and the `content` fields on an `AnswerDocument`.
+-   Searches within AAQ threads are possible by searching over the `question_content` fields on a `QuestionDocument`,
+    and the `content` fields on an `AnswerDocument`.
 
-* Searches can be limited to Question content and the content of Answers which are solutions by filtering by `is_solved: True`,
-and searching over `question_content` and `content` fields.
+-   Searches can be limited to Question content and the content of Answers which are solutions by filtering by `is_solved: True`,
+    and searching over `question_content` and `content` fields.
 
-* Aggregations on Answers,
-such as gathering contribution metrics,
-can be performed by aggregating `AnswerDocument`s.
+-   Aggregations on Answers,
+    such as gathering contribution metrics,
+    can be performed by aggregating `AnswerDocument`s.
 
 There may be some Question data indexed in `AnswerDocument`s which we never use.
