+++ kitsune/community/tests/test_templates.py
@@ -3,7 +3,7 @@ from pyquery import PyQuery as pq
 
 from kitsune.forums.tests import ThreadFactory
 from kitsune.questions.tests import AnswerFactory
-from kitsune.search.v2.tests import Elastic7TestCase
+from kitsune.search.tests import Elastic7TestCase
 from kitsune.sumo.templatetags.jinja_helpers import urlparams
 from kitsune.sumo.tests import LocalizingClient
 from kitsune.sumo.urlresolvers import reverse
