+++ kitsune/community/tests/test_utils.py
@@ -9,7 +9,7 @@ from kitsune.community.utils import (
 )
 from kitsune.products.tests import ProductFactory
 from kitsune.questions.tests import AnswerFactory
-from kitsune.search.v2.tests import Elastic7TestCase
+from kitsune.search.tests import Elastic7TestCase
 from kitsune.sumo.tests import LocalizingClient
 from kitsune.wiki.tests import DocumentFactory, RevisionFactory
