+++ kitsune/questions/management/commands/auto_archive_old_questions.py
@@ -7,7 +7,7 @@ from django.db import connection, transaction
 
 from kitsune.questions.models import Question, Answer
 
-from kitsune.search.v2.es7_utils import index_objects_bulk
+from kitsune.search.es7_utils import index_objects_bulk
 
 log = logging.getLogger("k.cron")
