+++ kitsune/questions/tests/test_templates.py
@@ -25,7 +25,7 @@ from kitsune.questions.tests import (
     tags_eq,
 )
 from kitsune.questions.views import NO_TAG, UNAPPROVED_TAG
-from kitsune.search.v2.tests import Elastic7TestCase
+from kitsune.search.tests import Elastic7TestCase
 from kitsune.sumo.templatetags.jinja_helpers import urlparams
 from kitsune.sumo.tests import (
     LocalizingClient,
