+++ kitsune/search/__init__.py
@@ -10,3 +10,5 @@ HIGHLIGHT_TAG = "strong"
 SNIPPET_LENGTH = 500
 
 NO_MATCH = _lazy("No pages matched the search criteria")
+
+default_app_config = "kitsune.search.apps.SearchConfig"
