+++ kitsune/search/decorators.py
