+++ kitsune/search/signals/questions.py
@@ -1,11 +1,11 @@
 from django.db.models.signals import post_save, post_delete, m2m_changed
-from kitsune.search.v2.es7_utils import (
+from kitsune.search.es7_utils import (
     index_object,
     delete_object,
     index_objects_bulk,
     remove_from_field,
 )
-from kitsune.search.v2.decorators import search_receiver
+from kitsune.search.decorators import search_receiver
 from kitsune.questions.models import Question, QuestionVote, Answer, AnswerVote
 from taggit.models import Tag
