+++ kitsune/search/signals/wiki.py
@@ -1,8 +1,8 @@
 from django.db.models.signals import m2m_changed, post_delete, post_save
 
 from kitsune.products.models import Product, Topic
-from kitsune.search.v2.decorators import search_receiver
-from kitsune.search.v2.es7_utils import delete_object, index_object, remove_from_field
+from kitsune.search.decorators import search_receiver
+from kitsune.search.es7_utils import delete_object, index_object, remove_from_field
 from kitsune.wiki.models import Document
