+++ kitsune/search/tests/test_es7_utils.py
@@ -1,12 +1,12 @@
-from kitsune.search.v2.tests import Elastic7TestCase
+from kitsune.search.tests import Elastic7TestCase
 from kitsune.questions.tests import QuestionFactory
 from django.test.utils import override_settings
-from kitsune.search.v2.es7_utils import index_objects_bulk
-from elasticsearch7.helpers.errors import BulkIndexError
-from elasticsearch7.exceptions import NotFoundError
-from kitsune.search.v2.documents import QuestionDocument
+from kitsune.search.es7_utils import index_objects_bulk
+from elasticsearch.helpers.errors import BulkIndexError
+from elasticsearch.exceptions import NotFoundError
+from kitsune.search.documents import QuestionDocument
 from unittest.mock import patch
-from kitsune.search.v2.base import SumoDocument
+from kitsune.search.base import SumoDocument
 
 
 @override_settings(ES_LIVE_INDEXING=False)
@@ -15,7 +15,7 @@ class IndexObjectsBulkTestCase(Elastic7TestCase):
         q_id = QuestionFactory(is_spam=True).id
         index_objects_bulk("QuestionDocument", [q_id])
 
-    @patch("kitsune.search.v2.documents.QuestionDocument.to_action", autospec=True)
+    @patch("kitsune.search.documents.QuestionDocument.to_action", autospec=True)
     def test_errors_are_raised_after_all_chunks_are_sent(self, mock_to_action):
         es_exception_will_be_thrown = False
         id_without_exception = None
