+++ requirements/default.in
@@ -36,8 +36,8 @@ django-user-agents~=0.4.0
 django-waffle~=2.1.0
 django-watchman~=1.1.1
 djangorestframework~=3.11.0
-elasticsearch~=1.2.0  # Listed explicitly because elasticutils does not constrain appropriately.
-elasticutils~=0.10.3
+elasticsearch~=7.10.0
+elasticsearch-dsl~=7.3.0
 factory-boy~=2.12.0
 Faker~=4.1.0
 google-api-python-client~=1.8.4
