+++ requirements/default.txt
@@ -1,4 +1,4 @@
-# SHA1:9f8b027f8fd160a4917a8b2dad728d19dc636f0c
+# SHA1:bb28c793188d4cd4376ba6fe948da67b3d570997
 #
 # This file is autogenerated by pip-compile-multi
 # To update, run:
@@ -59,6 +59,7 @@ certifi==2020.12.5 \
     --hash=sha256:1a4995114262bffbc2413b159f2a1a480c969de6e6eb13ee966d470af86af59c \
     --hash=sha256:719a74fb9e33b9bd44cc7f3a8d94bc35e4049deebe19ba7d8e108280cfd59830
     # via
+    #   elasticsearch
     #   requests
     #   sentry-sdk
 cffi==1.14.5 \
@@ -283,16 +284,16 @@ djangorestframework==3.11.2 \
     --hash=sha256:5cc724dc4b076463497837269107e1995b1fbc917468d1b92d188fd1af9ea789 \
     --hash=sha256:a5967b68a04e0d97d10f4df228e30f5a2d82ba63b9d03e1759f84993b7bf1b53
     # via -r requirements/default.in
-elasticsearch==1.2.0 \
-    --hash=sha256:53a822b01b0494df9658c04e87216180e5ecbe85b178673dd848b1f809d5c100 \
-    --hash=sha256:bb93d85ca7e50d68e4197852bb6dba6098bb772500f82f5db2ff4a80711505c9
+elasticsearch-dsl==7.3.0 \
+    --hash=sha256:0ed75f6ff037e36b2397a8e92cae0ddde79b83adc70a154b8946064cb62f7301 \
+    --hash=sha256:9390d8e5cf82ebad3505e7f656e407259cf703f5a4035f211cef454127672572
+    # via -r requirements/default.in
+elasticsearch==7.10.1 \
+    --hash=sha256:4ebd34fd223b31c99d9f3b6b6236d3ac18b3046191a37231e8235b06ae7db955 \
+    --hash=sha256:a725dd923d349ca0652cf95d6ce23d952e2153740cf4ab6daf4a2d804feeed48
     # via
     #   -r requirements/default.in
-    #   elasticutils
-elasticutils==0.10.3 \
-    --hash=sha256:569037ff11c3c11638f3d06857033424525ad54618bffb012bd43c32c50fbe3b \
-    --hash=sha256:88f2b91b8a498e0c903e4e70afb1fbae74d23f62f98bbe968c22154157cbe955
-    # via -r requirements/default.in
+    #   elasticsearch-dsl
 factory-boy==2.12.0 \
     --hash=sha256:728df59b372c9588b83153facf26d3d28947fc750e8e3c95cefa9bed0e6394ee \
     --hash=sha256:faf48d608a1735f0d0a3c9cbf536d64f9132b547dae7ba452c4d99a79e84a370
@@ -650,6 +651,7 @@ python-dateutil==2.8.1 \
     # via
     #   -r requirements/default.in
     #   botocore
+    #   elasticsearch-dsl
     #   faker
 python-decouple==3.4 \
     --hash=sha256:2e5adb0263a4f963b58d7407c4760a2465d464ee212d733e2a2c179e54c08d8f \
@@ -763,7 +765,7 @@ six==1.15.0 \
     #   django-extensions
     #   django-jsonfield
     #   django-statici18n
-    #   elasticutils
+    #   elasticsearch-dsl
     #   google-api-core
     #   google-api-python-client
     #   google-auth
